Modularized main.py to be divided into services
